Improve code separation
